requirements.txt
streamlit
openai-whisper
ffmpeg-python
numpy
streamlit-webrtc
av
